modulize url checking functions
